adblib: Add support for trackDevices with retry
Implementation of `AdbLibSession.trackDevices` to allow tracking
devices as long as an `AdbLibSession` is active, retrying connections
to ADB server under the hood if the ADB server becomes temporarily
unavailable.

* Also adds an in-memory cache (`SessionCache`) to AdbLibSession
  for storing generic (key, value) pairs having the same lifecycle as
  an `AdbLibSession` instance.

* Also adds a `AdbLibSession.scope` CoroutineScope to allow running
  coroutines with the same scope/lifecycle as an `AdbLibSession`
  instance.

To test the "retry" behavior of the track devices flow, this change
also adds the ability to restart `FakeAdbServer`, simulating
ADB server being stopped and restarted.

Bug: 188067646
Test: Included
Change-Id: I1d5eeafece02cac59c05bf148f18ff4cfdad6e75