adblib: Re-implement hearbeat for shell command
Change the implementation of `withCommandIdleTimeout` to measure actual
elapsed time instead of solely relying on `withTimeout`. The issue
with `withTimeout` is that, under heavy system load, a TimeoutException
could be thrown even though the shell command is active, merely because
the heartbeat detection coroutine can be "starved". The new
implementation measure actual elapsed time between heartbeats, meaning
the worst case scenario is that a `TimeoutException` can be thrown a
little too late instead of too eagerly (which is a better outcome in
case of heavy load).

Also, the implementation now uses a `StateFlow` instead of a
`Channel` to communicate between the hearbeat detector coroutine and
the shell command output. This improves performance, as a StateFlow
is essentially an `AtomicReference`, whereas a `Channel` synchronize
sender and receiver when using the "rendez vous" policy. This could
slow down the shell output process fs the hearbeat detection coroutine
is delayed due to heavy load.

This new implementation was testing by running

bazel test --runs_per_test 5000 //tools/base/adblib:adblib.tests.test

We used to get about 10 failures with 5,000 runs. We get none after
the fix.

Bug: 188067646
Test: Existing
Change-Id: I777129f7e2040625fe02b6d06b6f308868155955