adblib: Prevent allocations in `AdbInputChannel.read`
Returning an `Int` from a `suspend` function can lead to
allocation due to boxing of the return value (coroutines
always return an `Object` instance).

In the case of `AdbInputChannel.read`, we want to return
the number of bytes read. It turns out this return value can
be inferred by looking at how many bytes were stored in the
`ByteBuffer` after the `read` operation completes.

In this change, we rename `AdbInputChannel.read` to
`AdbInputChannel.readBuffer` and remove the `Int` return
value, and we create an `inline` extension function
`AdbInputChannel.read` that looks at the `ByteBuffer`
remaining bytes before and after `readBuffer` to infer
the return value (i.e. number of bytes read).

This ensures all `Int` allocations from calling
`AdbInputChannel.read` are removed.

Bug: 262404780
Test: Existing
Change-Id: Ib2af539a077aee9de10d641b0ae90e39f4ba20f3