Recover from exceptions during claim
If the device goes offline during claim, then even the
DefaultProvisionerPlugin will be unable to claim it. This results in an
exception that kills the DeviceProvisioner scope.

Instead, if no device is able to claim, log the condition, and wait for
the device to come online again. Also, if the device switches from
online to offline during the claim process, cancel the claim process,
and wait for the device to come online again.

Loosen up the visibility of FakeAdbSession / FakeAdbLoggerFactory so
that we can inspect adblib logs in tests. This allows us to observe and
verify the state of the device provisioner.

Bug: 309651672
Test: DeviceProvisionerTest (1000 runs)
Change-Id: I952926f9153ef0705f626787c7fc9bdfd071da9f