adblib: fix test logger delta instant handling
* Instead of keeping track of delta instant per logger instance,
we now centralize this to the (logger) factory so that reading
entries of a log is easier to comprehend: each delta instant
value in a log entry is a delta from the previous line instead
of a delta from the previous line of a given logger instance.

* Improve tracking so that we display delta instants in seconds
when the number of milliseconds is >= 1000.

All in all, this improve readability of these delta instant
values, making them more useful when they are needed.

Bug: n/a
Test: n/a (part of test framework)
Change-Id: I89817f329345c630b23a3544ed05eff89193a979